+++ PhilterDesktop.Tests/CsvRedactorTests.cs
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+using Phileas.Services.Office;
 
 using System.Text;
 using Phileas.Policy;