+++ PhilterDesktop/CsvRedactor.cs
@@ -19,6 +19,7 @@ using System.Text;
 using CsvHelper;
 using CsvHelper.Configuration;
 using Phileas.Model;
+using Phileas.Services.Office;
 using PhilterData;
 
 namespace PhilterDesktop