+++ PhilterDesktop/ModifyRedactionForm.cs
@@ -17,6 +17,7 @@
 using System.Diagnostics;
 using LiteDB;
 using Phileas.Policy;
+using Phileas.Services.Office;
 using PhilterData;
 using PhileasPolicy = Phileas.Policy.Policy;